Rewrite get status
