+++ love_letter/usecase/get_status.py
@@ -1,26 +1,11 @@
-from dataclasses import dataclass
-
 from love_letter.models import Game
-from love_letter.models.event import DomainEvent
+from love_letter.models.event import GetStatusEvent
 from love_letter.usecase.common import Presenter, game_repository
 
 
 class GetStatusInput:
     game_id: str
     player_id: str
-
-
-@dataclass
-class GetStatusEvent(DomainEvent):
-    game: Game
-
-
-class GetStatusPresenter(Presenter):
-    def as_view_model(self) -> Game:
-        for event in self.events:
-            if isinstance(event, GetStatusEvent):
-                return event.game
-        raise BaseException("Game is unavailable.")
 
 
 class GetStatus:
@@ -34,7 +19,3 @@ class GetStatus:
         input.game_id = game_id
         input.player_id = player_id
         return input
-
-    @classmethod
-    def presenter(cls) -> "GetStatusPresenter":
-        return GetStatusPresenter()
